+++ packages/no-modal/src/base/connector/utils.ts
@@ -1,5 +1,7 @@
 import { jwtDecode } from "jwt-decode";
 
+import type { IdentityTokenInfo } from "./interfaces";
+
 export function storageAvailable(type: "sessionStorage" | "localStorage"): boolean {
   let storageExists = false;
   let storageLength = 0;
@@ -60,6 +62,19 @@ export const saveToken = (userAddress: string, issuer: string, token: string) =>
 export const clearToken = (userAddress: string, issuer: string) => {
   if (storageAvailable("localStorage")) {
     return localStorage.removeItem(`${userAddress.toLowerCase()}_${issuer}`);
+  }
+  return null;
+};
+
+export const getCachedTokenInfo = (account: string, connectorName: string): IdentityTokenInfo | null => {
+  const saved = getSavedToken(account, connectorName);
+  if (!saved) return null;
+
+  try {
+    const parsed = JSON.parse(saved) as IdentityTokenInfo;
+    if (parsed.idToken && !checkIfTokenIsExpired(parsed.idToken)) return parsed;
+  } catch {
+    if (!checkIfTokenIsExpired(saved)) return { idToken: saved };
   }
   return null;
 };
